requirement.txt
argparse
yara
capstone
r2pipe
elftools
pefile
progressbar